fix tests, temporarily disabling some tests that need to be re-written
