+++ bigchaindb/pipelines/election.py
@@ -35,8 +35,7 @@ class Election:
         next_block = self.bigchain.get_block(
             next_vote['vote']['voting_for_block'])
 
-        block_status = self.bigchain.block_election_status(next_block['id'],
-                                                           next_block['block']['voters'])
+        block_status = self.bigchain.block_election_status(next_block)
         if block_status == self.bigchain.BLOCK_INVALID:
             return Block.from_dict(next_block)
