+++ bigchaindb/pipelines/vote.py
@@ -57,8 +57,7 @@ class Vote:
                                                    [([self.bigchain.me], 1)])
 
     def validate_block(self, block):
-        if not self.bigchain.has_previous_vote(block['id'],
-                                               block['block']['voters']):
+        if not self.bigchain.has_previous_vote(block['id']):
             try:
                 block = Block.from_dict(block)
             except (exceptions.InvalidHash):
