+++ tests/db/test_bigchain_api.py
@@ -82,12 +82,12 @@ class TestBigchainApi(object):
         block = b.create_block([tx])
         b.write_block(block)
 
-        assert b.has_previous_vote(block.id, block.voters) is False
+        assert b.has_previous_vote(block.id) is False
 
         vote = b.vote(block.id, b.get_last_voted_block().id, True)
         b.write_vote(vote)
 
-        assert b.has_previous_vote(block.id, block.voters) is True
+        assert b.has_previous_vote(block.id) is True
 
     @pytest.mark.genesis
     def test_get_spent_with_double_inclusion_detected(self, b, monkeypatch):
@@ -429,43 +429,6 @@ class TestBigchainApi(object):
         assert retrieved_block_1 == retrieved_block_2
 
     @pytest.mark.genesis
-    def test_more_votes_than_voters(self, b):
-        from bigchaindb.common.exceptions import MultipleVotesError
-
-        block_1 = dummy_block()
-        b.write_block(block_1)
-        # insert duplicate votes
-        vote_1 = b.vote(block_1.id, b.get_last_voted_block().id, True)
-        vote_2 = b.vote(block_1.id, b.get_last_voted_block().id, True)
-        vote_2['node_pubkey'] = 'aaaaaaa'
-        b.write_vote(vote_1)
-        b.write_vote(vote_2)
-
-        with pytest.raises(MultipleVotesError) as excinfo:
-            b.block_election_status(block_1.id, block_1.voters)
-        assert excinfo.value.args[0] == 'Block {block_id} has {n_votes} votes cast, but only {n_voters} voters'\
-            .format(block_id=block_1.id, n_votes=str(2), n_voters=str(1))
-
-    def test_multiple_votes_single_node(self, b, genesis_block):
-        from bigchaindb.common.exceptions import MultipleVotesError
-
-        block_1 = dummy_block()
-        b.write_block(block_1)
-        # insert duplicate votes
-        for i in range(2):
-            b.write_vote(b.vote(block_1.id, genesis_block.id, True))
-
-        with pytest.raises(MultipleVotesError) as excinfo:
-            b.block_election_status(block_1.id, block_1.voters)
-        assert excinfo.value.args[0] == 'Block {block_id} has multiple votes ({n_votes}) from voting node {node_id}'\
-            .format(block_id=block_1.id, n_votes=str(2), node_id=b.me)
-
-        with pytest.raises(MultipleVotesError) as excinfo:
-            b.has_previous_vote(block_1.id, block_1.voters)
-        assert excinfo.value.args[0] == 'Block {block_id} has {n_votes} votes from public key {me}'\
-            .format(block_id=block_1.id, n_votes=str(2), me=b.me)
-
-    @pytest.mark.genesis
     def test_improper_vote_error(selfs, b):
         from bigchaindb.common.exceptions import ImproperVoteError
 
@@ -476,7 +439,7 @@ class TestBigchainApi(object):
         vote_1['signature'] = 'a' * 87
         b.write_vote(vote_1)
         with pytest.raises(ImproperVoteError) as excinfo:
-            b.has_previous_vote(block_1.id, block_1.id)
+            b.has_previous_vote(block_1.id)
         assert excinfo.value.args[0] == 'Block {block_id} already has an incorrectly signed ' \
                                         'vote from public key {me}'.format(block_id=block_1.id, me=b.me)
